Fix test errors
