+++ package.json
@@ -80,9 +80,9 @@
     "onCommand:openshift.service.describe",
     "onCommand:openshift.service.describe.palette",
     "onCommand:openshift.component.folder.create",
+    "onCommand:openshift.explorer.reportIssue",
     "onCommand:clusters.openshift.openProjectConsole",
-    "onCommand:clusters.openshift.useProject",
-    "onCommand:openshift.issueTracker"
+    "onCommand:clusters.openshift.useProject"
   ],
   "main": "./out/src/extension",
   "contributes": {
@@ -131,7 +131,7 @@
         }
       },
       {
-        "command": "openshift.issueTracker",
+        "command": "openshift.explorer.reportIssue",
         "title": "Report Extension Issue on GitHub",
         "category": "OpenShift",
         "icon": {
@@ -511,7 +511,7 @@
           "group": "navigation"
         },
         {
-          "command": "openshift.issueTracker",
+          "command": "openshift.explorer.reportIssue",
           "when": "view == openshiftProjectExplorer",
           "group": "navigation"
         }
