+++ test/openshift/cluster.test.ts
@@ -25,10 +25,10 @@ const keytar: any = getVscodeModule('keytar');
 
 suite('Openshift/Cluster', () => {
     let sandbox: sinon.SinonSandbox;
-    let execStub: sinon.SinonStub, 
-        commandStub: sinon.SinonStub, 
+    let execStub: sinon.SinonStub,
+        commandStub: sinon.SinonStub,
         inputStub: sinon.SinonStub,
-        infoStub: sinon.SinonStub, 
+        infoStub: sinon.SinonStub,
         loginStub: sinon.SinonStub,
         quickPickStub: sinon.SinonStub;
     const testData: CliExitData = {
